Send files with post request
Now the form can send files using a POST request with fetch, and the server sends a response with them.

index.js is now app.js

Changed app port from 8080 to 3737

To avoid collisions, now in app.html the name/id "file-load" is "fileLoad", and the name/id "export" is "output".