Form number input improved and server errors fixed
Now the form inputs of type number take numbers and backspace only, the rest of the keys don't work.

Now the form inputs of type range and number are linked, so when one of the two changes, the other changes too correspondingly.

Now the server handles files properly. Before it would take a file and sometimes either it didn't finish writing the file properly or it deleted the file before exporting it, turning in an empty zip file, or with corrupted images that couldn't display properly.