PNG quality and self-explanatory variable names
The problem that didn't allow the proper compression of PNG files has been finally addressed.

To make the code more readable, the functions that convert the files and the variables that they take now have self-explanatory names.

Added some script commands in the package.json to make the building of the code and the start of the web app easier.